add integration test

+++ src/main/java/com/capgemini/training/api/service/DeleteCustomerDetailsService.java
@@ -1,6 +1,7 @@
 package com.capgemini.training.api.service;
 
 import com.capgemini.training.api.repository.CustomerRepository;
+import org.apache.logging.log4j.util.Strings;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,7 +13,10 @@ public class DeleteCustomerDetailsService {
     this.customerRepository = customerRepository;
   }
 
-  public void deleteCustomerDetails(String customerId){
+  public void deleteCustomerDetails(String customerId) {
+    if (Strings.isEmpty(customerId)) {
+      throw new IllegalArgumentException();
+    }
     customerRepository.deleteById(customerId);
   }
 }
